% Generated by build script
\name{osv_query}
\alias{osv_query}
\title{Query OSV for vulnerabilities affecting a package}
\usage{
osv_query()
}
\arguments{
\item{package}{Package name (e.g. "lodash", "requests", "log4j-core")}
\item{ecosystem}{Package ecosystem: "npm", "PyPI", "crates.io", "Maven",}
}
\value{tibble: id (character), summary (character), published (character),}
\description{Query OSV for vulnerabilities affecting a package}
